Collect AVPlayerItem-related code
